move createSession again
