+++ app/session/handlers.ts
@@ -1,7 +1,7 @@
 import { execSync } from "child_process";
 import type { Client, Message } from "whatsapp-web.js";
 import { getAiResponse } from "../langchain";
-import { PERMANENT_FAILURE_REASONS, RECONNECT_CONFIG } from ".";
+import { createSession, PERMANENT_FAILURE_REASONS, RECONNECT_CONFIG } from ".";
 import { sessions } from ".";
 import { raceWithTimeout } from "../utils";
 
@@ -109,7 +109,6 @@ export async function scheduleReconnect(companySlug: string, reason: string): Pr
             await safeDestroyClient(companySlug);
 
             console.log(`[RECONNECT] ${companySlug}: criando sessao nova...`);
-            const { createSession } = await import("./status");
             await createSession(companySlug);
 
             await new Promise<void>((resolve) => setTimeout(resolve, 15000));
